change ids to Longs for account and profile.

+++ spring-nanotrader-asynch-services/src/test/java/org/springframework/nanotrader/asynch/service/TradingServiceAsynchTests.java
@@ -62,7 +62,7 @@ public class TradingServiceAsynchTests {
 
 	@Test
 	public void testService() {
-		Integer accountId = accountDataOnDemand.getRandomAccount().getAccountid();
+		Long accountId = accountDataOnDemand.getRandomAccount().getAccountid();
 		org.springframework.nanotrader.data.domain.Quote dquote = quoteService.findBySymbol("GOOG");
 		Quote quote = new Quote();
 		mapper.map(dquote, quote);
