+++ spring-nanotrader-data/src/main/java/org/springframework/nanotrader/data/repository/HoldingAggregateRepositoryImpl.java
@@ -54,7 +54,7 @@ public class HoldingAggregateRepositoryImpl implements HoldingAggregateRepositor
 	}
 
 	@Override
-	public HoldingSummary findHoldingAggregated(Integer accountId) {
+	public HoldingSummary findHoldingAggregated(Long accountId) {
 	
 		HoldingSummary holdingSummary = new HoldingSummary();
 		List<HoldingAggregate> holdingRollups = new ArrayList<HoldingAggregate>();
