+++ spring-nanotrader-service-support/src/main/java/org/springframework/nanotrader/service/domain/Account.java
@@ -83,13 +83,13 @@ public class Account implements Serializable {
     }
 
 
-    private Integer accountid;
+    private Long accountid;
 
-	public Integer getAccountid() {
+	public Long getAccountid() {
         return this.accountid;
     }
 
-	public void setAccountid(Integer id) {
+	public void setAccountid(Long id) {
         this.accountid = id;
     }
